Market data parsing
Core features for parsing individual markets has been laid down. They're
still split into a few sections, and for now are using text fields
instead of variables in order to observe the results. Need to implement
some cleanup handlers to get rid of whitespace and invalid markets such
as "BTC-BTC" and "LTC-LTC" and find an easy way to cache which currency
pairs are valid without having to query the API every time. Added
options to destroy data on close.